refactor: App으로 로직 분리
