+++ src/dom/constants.js
@@ -8,11 +8,13 @@ export const ID_MAP = {
     showResult: 'show-result',
   },
   container: 'container',
+  app: 'app',
 };
 
 export const CLASS_NAME_MAP = {
   modal: 'modal',
   overlay: 'overlay',
+  price: 'price',
   winningNumber: 'winning-number',
   winningBox: 'winning-box',
   winningNumberBox: 'winning-number-box',
